Use config gem to store configuration
This is the pattern we use across DLSS codebases, so this is both for consistency and for a more robust pattern going forward. I wanted this gem around for two bits of past work (adding BCP47 language code validation and fixing field flattening in the cardinality adjuster). I decided to leave those PRs without this change, opting to put this in its own PR.

This PR also includes:
* Use `config` gem to express which fields need flattening in the cardinality adjuster
* Load configuration from `{gem_root}/settings.yml`
* Sort gems alphabetically in `Dlme::CLI` class to avoid suggesting the requires are ordered in some way
* Remove method (`acceptable_bcp47_codes`) that was playing the role of the `config` gem
* Technical debt clean-up: remove `LineWriterFixMixin` and `DlmeDebugWriter` which are no longer necessary in Traject >= 3.1.0.