Rejigger how requires are used
This fixes an issue we have missed thus far w/ the `singleton` library not being required anywhere. The changes in this branch concentrate most requires for the gem into `lib/cli.rb` since it is the only entrypoint into the gem at this point. We require traject and traject_plus there so we do not need to require it higgledy-piggledy throughout the codebase. We require `singleton` in the `RecordCounter` class since it is only used in that one context. And we require `cli` in the specs so that the test suite uses the libraries *as required in the gem*, which helps surface errors specifying requirements.