Check whether *any* exception is raised in integration spec
This is currently failing because of a data error, and indeed, there could be several dozen or hundred more such errors. As we fix data that does not cleanly map, we should revisit this PR until we get a clean build. As to why we make this change? Because only checking for `RuntimeError` allows for false positives, where we may be missing exceptions.