Use YAML to parse
We are having some edge case problems parsing serialized Python
lists as CSV. After reading [this] I thought it could be interesting to
try to parse the lists as YAML rather than CSV.

```
irb(main):010:0> YAML.load(%q(['123', '456']))
=> ["123", "456"]
irb(main):011:0> YAML.load(%q(['abc', "'def,' 123"]))
=> ["abc", "'def,' 123"]
```

The second one here is effectively what our current CSV parsing is choking on.

A nice side effect of parsing as YAML is that we can remove some of the
initial string manipulation that we needed for parsing as CSV,
since spaces don't need to be removed between strings, and we can leave
the square brackets in place.

[this]: https://stackoverflow.com/questions/4477127/ruby-parsing-a-string-representation-of-nested-arrays-into-an-array/17271822#17271822